front/back/requirement.txt
Flask
mysql-connector-python
bcrypt